feat: Add `standalone-wasi` feature
This commit adds a `standalone-wasi` feature to the core crate.

This is done as a preliminary improvement to correctly support unit tests
in the core crate through cargo wasi.

This change allows conditionally compiling abi imports for data passing, doing so
doesn't pollute the test runs with imports that are not needed.