feat: Support windows
This commit ensures Windows support.

As of rust 1.3x fs::canonicalize  returns Windows NT UNC paths
(\\?\C:\foo) which are not gracefully handled by all windows versions.

This commit swaps the usage of files  and instead passes the contents
of the javascript file to the wizer init process through stdin, ensuring that files
are skipped altogether.