handle arbitrary-sized closures correctly in `new_callback`
Previously, `Context::new_callback` would not work reliably for closures larger
than 4 bytes in size.  This fixes that by moving the closure into a `Box` and
passing the pointer to QuickJS instead of the closure itself.

Note that the boxed closure is leaked, which seems unavoidable given that
QuickJS provides no way to specify custom finalizers for `JSValue`s.  However,
`quickjs-wasm-rs` already leaks memory in various ways, which we consider
acceptable since it is intended for short-lived programs.

This also adds a `'static` bound to the closure type, ensuring that non-`static`
borrows are rejected at compile time, which makes `new_callback` safe to use
without an `unsafe` block (although `unsafe` will generally be needed in the
_body_ of the closure -- use `wrap_callback` to avoid that).

Signed-off-by: Joel Dice <joel.dice@fermyon.com>