Switch to `rquickjs` (#618)
* chore: Initial migration to rquickjs

* More progress on migration:

Migrated the `Console` and `Random` APIs; got to the `StreamIO` API.

* Migrate all of the APIs to use rquickjs

This commit migrates the `console`, `stream_io`, `random` and `text_encoding`  APIs to use rquickjs. 

One notable change in this commit is the introduction of the `Args` struct to tie the lifetime of `Ctx<'js>` and `Rest<Value<'js>>` arguments, given that explicit lifetime binding is not possible in Rust closures at the moment.

* Finalize migration of the toolchain as a whole

* Uncomment integration tests

* Throw `TypeError` when `fatal` decoding is specified.

* Fix clippy errors

* Update documentation where applicable and remove reference to older crates

* Handle converting from JS strings to Rust strings for text encoding

* Migrate serializers and deserializers to use rquickjs and move them under the javy crate

* Move quickcheck to the top level Cargo.toml 


 `cargo` doesn't seem to pickup the fact that `quickcheck` is only used in tests.

* Update versions and CHANGELOGs

* Audit dependencies

And prune unused ones

* Improve safety comment

* Port previous implementation of read sync

I wanted to avoid as much as possible using the underlying QuickJS unsafe APIs, but that introduced behavioral changes and bugs. So for now, I'm sticking to the previous implementation.

* Review comments

* Improve comment for why `ManuallyDrop`

* Use `ManuallyDrop` for Context

* Format comments where applicable

* Use rquickjs 0.6.0 through a fork

This commit introduces rquickjs 0.6.0 through a fork that contains Wasm specific performance improvements. We intend to upstream these improvements.

* Cargo vet config

Adds special policy for rquickjs given that we're currently in a fork.

* Remove comment, cargo vet doesn't like them