docs: Improve READMEs for `apis` and `javy` crates (#640)
* docs: Slight improvement to `javy-apis`

In preparation to adding more APIs I decided to do a very light cleanup of some of the documentation in this crate. I opted to remove the documentation in the README to reduce the overhead of having to maintain the same documentation in multiple places.

* Review comments

* Bring example to README and fix error

Brings the APIs example to the README and fixes a dangling statement.

* Use the right example

* Standardize the READMEs even more

Adds crates.io and docs.rs badges and updates the  `javy` crate with similar updates. 

Includes a link to the documentation to ensure that there's a direct navigation from the README to the crate-level docs.