Test the JSON.{parse, stringify} implementation
This commit adds `javy-test-macros` as a development dependency in the
`javy` crate and adds 262 test for Javy's custom implementation of
`JSON.{parse,stringigy}`.

The 262 suite is pulled in as a submodule.

In order to test the custom implementation, this commit explicitly calls
`config.override_json_parse_and_stringify` when creating the default
runtime used in Javy. The override configuration was introduced in
https://github.com/bytecodealliance/javy/pull/651.