Introduce the `javy-config` crate
This commit prepares the terrain to bridge the gap in configurability
between the CLI and Javy, the crate.
The intention behind introducing a new crate is to:

* Reduce code duplication and sync the options between the CLI and the
  core crate.
* Make it easy to pass the options into WebAssemlby by using bitflags.

This PR doesn't introduce any new functionality. A follow up PR will
include new commands in the CLI which will make use of the share
configuration.