Rename test-macros and config crates
This commit is a minor clean-up to rename the `test-macros` and `config`
crates directories so that they adhere to the naming convention of the rest of the
crates. This is simply a mechanical change.

Additionally, the `publish = false` property is added to crates that are
not mean to be published.