Improve cargo feature handling and crypto APIs definition
This commit improves the `crypto` feature handling by making the
relationship between crypto and the experimental_event_loop explicit. If
`crypto` is defined, the `experimental_event_loop` feature will also be
enbabled.

Additionally, this commit makes it easier to define the async crypto
APIs by introducing a `crypto.js` file that defines JS bits to make it
easier to return a promise.