CLI improvements
This commit is a first step toward
https://github.com/bytecodealliance/javy/issues/742.

Note that the migration is not totally done. Aside from putting in
together the initial infrastructure to use `ValueParserFactory` the main
intention of this change is to make it easier to derive argument name
and documentation from option groups, said differently, make the process
of adding new arguments more manageable and less repetitive.