Add support for `-J` option group
This commit adds support for the `-J` option group and enables
`javy-json` and `override-json-parse-and-stringify` by default.

Additionally, this change improves how the javy cli integration tests
are defined by introducing a `javy_cli_test` which:

* Produces a test per supported command (`compile` and `build`)
* Provides a set of defaults depending on the command under test

Generating a test per command improves the developer experience of
determining failures on particular command configurations.

Lastly, this change bumps the provider to v3, given that new
funcionality is introduced as part of enabling `javy-json` and
`override-json-parse-and-stringify`. In order to test that the `compile`
command remains "frozen", using version 2 of the provider, this commit
introduces a static copy of the v2 provided downloaded from the releases
section.