Fully thread the `source-compression` value
This commit fixes the handling of the `source-compression` option when
`C dynamic` is set. Even though the codegen option gropup defines the
options as `true` by default, it was not threaded correctly.

Additionally this commit set `source-compression` as `true` in the
default implementation of the builders to ensure that the CLI options
intent is accurately reflected.