Simplify CI & development make targets
This commit is an attempt to simplify and unify our build scripts,
with the idea of potentially making it easier to integrate new
crates in the future and more importantly avoid divergence between our
testing in CI and our testing locally. For instance, as part of this
change, I realized that at some point we stopped running `cargo fmt`
in CI, which is not ideal. `cargo fmt`  is not runnig as part of this
change yet though, I plan to add that as a follow-up.