make: Correctly format `lint-wasi-targets` target
The formatting error in the Makefile prevented the lint step from
completing and from reporting the lint errors declared in the
`wasi_p1` plugin api.

In CI, this looked like:

   --target=wasm32-wasip1 --all-targets --all-features -- -D warnings
   /bin/sh: 1: --target=wasm32-wasip1: not found
   make: [Makefile:13: lint-wasi-targets] Error 127 (ignored)

Ref: https://github.com/bytecodealliance/javy/actions/runs/19335192688/job/55308089241#step:4:21

Everything worked, given that the non-zero exit is ignored.
I'm currently not sure what's the best way to ensure that this doesn't
happen in the future.

Additionally, to make targets easier to read, this commit also divides
the p1/p2 targets, which will hopefully make this process less error-prone.